Implement loading condtional so it would render after the axios get request
